requirements.txt
livekit-agents~=1.3
livekit-plugins-deepgram~=1.3
livekit-plugins-google~=1.3
livekit-plugins-silero~=1.3
python-dotenv~=1.0
